Rename `*_unchecked()` to `*_maybe_unchecked()` and add debug assertions
When values of custom string types created, one of them must be called.
By adding assertions to all constructors, we can make it certain that
all values are validated at least once when they are created.

Note that `Self(v)` style value creation should not be used outside the
constructors, because we cannot run assertions automatically for such
expressions.